ChatProtocolAgent/arxiv_agent/requirements.txt
uagents==0.22.3
arxiv==2.1.3
google-generativeai==0.8.5
python-dotenv==1.0.1
uagents_core==0.2.3
